Switch from SF to OC for evrc, qcelp, raw aac, and LPA
external/opencore: adds usePVPlayer() function as a static
member of the PVPlayer class to
ask for an instance of PV_PLAYER instead of STAGEFRIGHT_PLAYER in the
following cases:
-qcelp and evrc files (not supported by stagefright)
-raw .aac files
-files with mp3 streams that are of sufficient length to use
opencore's LPA implementation
-aac streams in an mp4/3ggp container that are of sufficient length to
use opencore's LPA implementation

Returns OK if the file meets the above criteria;
returns UNKNOWN_ERROR otherwise.

Only switches to opencore for LPA if there is not another
instanceusing LPA already.

Change-Id: I95af225de19ccbc9f28e096a9655affdd1c7a8d7